requirements.txt
Flask
google-cloud-translate
python-dotenv
